Fix according to review
