+++ src/test/java/seedu/address/logic/parser/DeleteCommandParserTest.java
@@ -34,7 +34,9 @@ public class DeleteCommandParserTest {
 
     @Test
     public void parse_invalidMultipleArgs_throwsParseException() {
-        assertParseFailure(parser, "a b", String.format(MESSAGE_INVALID_COMMAND_FORMAT, DeleteCommand.MESSAGE_USAGE));
-        assertParseFailure(parser, "2030 s", String.format(MESSAGE_INVALID_COMMAND_FORMAT, DeleteCommand.MESSAGE_USAGE));
+        assertParseFailure(parser, "a b", String.format(MESSAGE_INVALID_COMMAND_FORMAT,
+                DeleteCommand.MESSAGE_USAGE));
+        assertParseFailure(parser, "2030 s", String.format(MESSAGE_INVALID_COMMAND_FORMAT,
+                DeleteCommand.MESSAGE_USAGE));
     }
 }
