+++ src/test/java/seedu/address/testutil/TestUtil.java
@@ -5,9 +5,6 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 
-import seedu.address.commons.core.index.Index;
-import seedu.address.model.Model;
-
 /**
  * A utility class for test cases.
  */
@@ -30,20 +27,6 @@ public class TestUtil {
         }
         return SANDBOX_FOLDER.resolve(fileName);
     }
-
-//    /**
-//     * Returns the middle index of the person in the {@code model}'s person list.
-//     */
-//    public static Index getMidIndex(Model model) {
-//        return Index.fromOneBased(model.getFilteredPersonList().size() / 2);
-//    }
-//
-//    /**
-//     * Returns the last index of the person in the {@code model}'s person list.
-//     */
-//    public static Index getLastIndex(Model model) {
-//        return Index.fromOneBased(model.getFilteredPersonList().size());
-//    }
 //
 //    /**
 //     * Returns the person in the {@code model}'s person list at {@code index}.
